requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
email-validator==2.2.0
python-dotenv==1.0.1
pymongo==4.8.0
werkzeug==3.0.4
passlib==1.7.4
Flask-Mail==0.9.1
